+++ platform/macos/src/MGLMapView.mm
@@ -272,14 +272,13 @@ public:
     _mbglView = new MGLMapViewImpl(self);
 
     // Delete the pre-offline ambient cache at
-    // ~/Library/Caches/com.mapbox.sdk.ios/cache.db.
+    // ~/Library/Caches/com.mapbox.MapboxGL/cache.db.
     NSURL *cachesDirectoryURL = [[NSFileManager defaultManager] URLForDirectory:NSCachesDirectory
                                                                        inDomain:NSUserDomainMask
                                                               appropriateForURL:nil
                                                                          create:NO
                                                                           error:nil];
-    cachesDirectoryURL = [cachesDirectoryURL URLByAppendingPathComponent:
-                          [NSBundle mgl_frameworkBundle].bundleIdentifier];
+    cachesDirectoryURL = [cachesDirectoryURL URLByAppendingPathComponent:@"com.mapbox.MapboxGL"];
     NSURL *legacyCacheURL = [cachesDirectoryURL URLByAppendingPathComponent:@"cache.db"];
     [[NSFileManager defaultManager] removeItemAtURL:legacyCacheURL error:NULL];